Reversed attack/damage logic
The attack is now triggered directly on the animals attack colliders, and for the bunny on the the collider of the poop projectile.

Each player controller toggles the collider locally which triggers the
OnTriggerEnter event in the PlayerAttack script attached to each
attack collider.

PlayerAttack handles attack between players and when players attack
NPCs, PlayerEffects still handles fall and firewall damage on players.

The old solution triggered more attack events than it should which had
strange side-effects like the fox being able to kill everyone with 2-3
bites, this should not happen now, as the events are only triggered
once per attack, and will not start if the attack animation is still in
effect.

Attack and healing/self-damage are handled differently in PlayerHealth:
- Attack is called from the attacker via the Attack method, which will notify the server, and
the server will tell the damaged player that they have been attacked,
which triggers the applyDamage method on the damaged client.
- Healing is called from an NPC for example, and the player applies the
healing directly to itself.
- TakeDamage is available when the damage can be applied directly to the
damaged player without going through the server, this is used for
example by the OnPoopGrenade method in PlayerEffects.

- host attack client: OK
- client attack host: OK
- client attack client: OK
- blood particle effect: OK
- damage: OK
- stats: OK
- die: OK
- sync body remove when die: OK
- disable attack when spectating: OK
- win: OK
- ranking: OK
- fox: OK
- bird: OK
- moose: OK
- bunny: OK
- all animals can kill NPCs (with host and client): OK
- abilities: OK
- fall: OK
- firewall: OK